Replace usage of deprecated alterConfigs with incrementalAlterConfigs where possible
